Nettoyage Js

+++ Front/js/produit.js
@@ -17,22 +17,24 @@ async function main() {
 }
 
 
+// Récupération de l'ourson choisi par l'utilisateur
 function getTeddy() {
   return fetch(urlTeddy)
   .then((response) => {
-    return response.json()
+    return response.json();
   })
   .then(function (teddy) {
-    return teddy
+    return teddy;
   })
   .catch((error) => {
     document.querySelector('#notification').innerHTML = `
     <div class="notification_alert" role="alert">
       ${error}
-    </div>`
+    </div>`;
   })
 }
 
+// Affichage de l'ourson choisi par l'utilisateur
 function displayTeddy(teddy) {
 
   // passage du prix en €
@@ -74,12 +76,9 @@ function displayTeddy(teddy) {
       // Au click sur le bouton validate
   document.querySelector('#validate').addEventListener('click', () => {
 
-      console.log(oldCart); 
-
     if (oldCart == null){
       // Si il est vide
         //  Récupérer les informations de l'ourson affiché est le mettre dans un tableau oursons
-        
         oursons.push(ourson);
 
         //  passer le tableau en JSON et le mettre dans le localStorage
@@ -97,88 +96,14 @@ function displayTeddy(teddy) {
       oursons.push(ourson);
 
       //  passer le tableau en JSON et le mettre dans le localStorage
-      localStorage.setItem('teddy', JSON.stringify(oursons));
+      
     }
-      alert(`${teddy.name} est rajouté au panier`)
-      document.documentElement.scrollTop = 0
-  })
+    
+    if (window.confirm(`Vous avez choisi ${teddy.name} \n Voulez confirmer ce choix`)) {
+        localStorage.setItem('teddy', JSON.stringify(oursons));
+        document.documentElement.scrollTop = 0;
+      }else{
+        location.href = 'index.html';
+      }
+  });
 }
-
-// Affichage de l'urlTeddy pour la demande au backend dans la console (pour vérification)
- 
-// Demande au backend les informations de l'ourson sélectionné par l'utilisateur
-// fetch(urlTeddy)
-//   .then(response =>
-//      response.json().then((data) => {
-//         // Affichage des informations de l'ourson sélectionné par l'utilisateur dans la console (pour vérification)
-//           console.log(data);
-
-//         // passage du prix en €
-//         let price = new Intl.NumberFormat('fr-FR', { style: 'currency', currency: 'EUR' }).format(data.price / 100);
-        
-//         // Insertion Name
-//         document.querySelector('.card_product-name').innerHTML = `<h3>${data.name}</h3> `;  
-
-//         // Insertion Photo  URL
-//         document.querySelector('.card_product-img').innerHTML = `<img src="${data.imageUrl}" alt= "Photo ${data.name}">`;
-
-//         // Insertion Description + Prix
-//         document.querySelector('.card_product-description').innerHTML = `
-//           <p>${data.description}</p>
-//           <p class="prix">${price}</p>`;
-        
-//         // Affichage des selections de couleurs possible sur la page html
-//         for (let i = 0; i < data.colors.length; i++) {
-//           document.querySelector('#select_color').innerHTML += `<option value="${data.colors[i]}">${data.colors[i]}</option>`;
-//         }
-
-//         // Récupération des informations à rajouter dans le panier
-//         let ourson = {
-//           id: data._id,
-//           nom_ourson: data.name,
-//           prix : price,
-//           couleur : data.colors, // Valeurs par default pour le MVP
-//           quantité : 1, // Valeurs par default pour le MVP
-//         };
-
-//         // Déclaration du tableau "oursons"
-//         let oursons = [];
-
-       
-//         // Afficher le LocalStorage
-//         let oldCart = JSON.parse(localStorage.getItem('teddy'));
-//         //  vérifier si le localStorage est vide
-//         // Au click sur le bouton validate
-//         document.querySelector('#validate').addEventListener('click', () => {
-
-//             console.log(oldCart); 
-
-//             if (oldCart == null)
-//             {
-//               // Si il est vide
-//                 //  Récupérer les informations de l'ourson affiché est le mettre dans un tableau oursons
-                
-//                 oursons.push(ourson);
-
-//                 //  passer le tableau en JSON et le mettre dans le localStorage
-//                 localStorage.setItem('teddy', JSON.stringify(oursons));
-//             } else {
-//               //  Si il n'est pas vide
-//                 //  rajouter les informations de l'ancien panier au tableau oursons
-//                 for (let i = 0 ; i < oldCart.length ; i++){
-//                   oursons.push(oldCart[i]);
-//                 }
-
-//                 //  Récupérer les informations de l'ourson affiché est le mettre dans un tableau oursons
-//                 oursons.push(ourson);
-
-//                 //  passer le tableau en JSON et le mettre dans le localStorage
-//                 localStorage.setItem('teddy', JSON.stringify(oursons));
-//             }
-
-//           
-
-//         })
-//       })
-//       .catch((err) => console.log('Erreur : ' + err))
-//     )
